Move constants to core directory
