+++ msgraph/cli/core/__init__.py
@@ -1,17 +1,13 @@
 import six
-import sys
 import types
 import pkgutil
 from knack import CLI
 from knack.cli import logger
-from knack.util import CLIError
-from knack import CLICommandsLoader
-from importlib import import_module
 from importlib import import_module
 from collections import OrderedDict
 from knack.arguments import ignore_type
 from knack.deprecation import Deprecated
-from knack import CLICommandsLoader, CLICommand, ArgumentsContext
+from knack import CLICommandsLoader, ArgumentsContext
 from knack.introspection import extract_args_from_signature, extract_full_summary_from_signature
 
 from msgraph.cli.core.commands._util import _load_module_command_loader, _load_extension_command_loader
@@ -22,7 +18,7 @@ from msgraph.cli.core.commands._util import get_arg_list
 from msgraph.cli.core.commands.client_factory import resolve_client_arg_name
 from msgraph.cli.core.commands.parameters import GraphArgumentContext
 from ._help import GraphCliHelp
-from .commands.constants import EXCLUDED_PARAMS
+from msgraph.cli.core.constants import EXCLUDED_PARAMS
 
 __version__ = '1.0.0'
